services/django/requirements.txt
Django==4.2.11
psycopg2-binary==2.9.9
redis==5.0.1
kafka-python==2.0.2
minio==7.2.3
requests==2.31.0
Pillow==10.2.0
djangorestframework==3.14.0
django-redis==5.4.0
celery==5.3.6
python-dotenv==1.0.1
gunicorn==21.2.0
whitenoise==6.6.0
django-widget-tweaks==1.5.0
pandas==2.2.0
plotly==5.18.0
